legal-doc-catalog/seed/requirements.txt
python-docx
supabase
python-dotenv
